+++ Cargo.toml
@@ -12,7 +12,8 @@ members = [
     "libs/fs",
     "libs/heap",
     "libs/interrupt",
-    "libs/ipc",
+    "libs/proc",
+    "libs/syscall",
     "libs/kernel_module",
     "libs/kernel_module/macros",
     "libs/log",
