+++ libs/kernel_module/Cargo.toml
@@ -12,7 +12,8 @@ edition = { workspace = true }
 log = { path = "../log" }
 memory = { path = "../memory" }
 kernel_module_macros = { path = "./macros" }
-ipc = { path = "../ipc" }
+syscall = { path = "../syscall" }
+proc = { path = "../proc" }
 anyhow = { workspace = true }
 spin = { workspace = true }
