+++ sophon/src/task/scheduler/mod.rs
@@ -1,12 +1,11 @@
 mod round_robin;
 
-use crate::arch::*;
 use alloc::sync::Arc;
 use atomic::{Atomic, Ordering};
 use core::fmt::Debug;
 use core::ops::Deref;
 
-use super::{task::Task, Message, TaskId};
+use super::{task::Task, TaskId};
 
 /**
  *                        ___________
@@ -42,56 +41,6 @@ pub trait AbstractScheduler: Sized + 'static {
 
     fn schedule(&self) -> !;
     fn timer_tick(&self);
-
-    fn unblock_sending_task(&self, id: TaskId, status: isize) {
-        let _guard = interrupt::uninterruptible();
-        let task = self.get_task_by_id(id).unwrap();
-        assert_eq!(
-            task.scheduler_state::<Self>().load(Ordering::SeqCst),
-            RunState::Sending
-        );
-        // Set response
-        task.context.set_response_status(status);
-        // Add this task to ready queue
-        self.mark_task_as_ready(task)
-    }
-
-    fn unblock_receiving_task(&self, id: TaskId, status: isize, m: Message) {
-        let task = self.get_task_by_id(id).unwrap();
-        assert_eq!(
-            task.scheduler_state::<Self>().load(Ordering::SeqCst),
-            RunState::Receiving
-        );
-        // Set response
-        task.context.set_response_message(m);
-        task.context.set_response_status(status);
-        // Add this task to ready queue
-        self.mark_task_as_ready(task)
-    }
-
-    fn block_current_task_as_sending(&self) -> ! {
-        let _guard = interrupt::uninterruptible();
-        let task = self.get_current_task().unwrap();
-        assert_eq!(
-            task.scheduler_state::<Self>().load(Ordering::SeqCst),
-            RunState::Running
-        );
-        task.scheduler_state::<Self>()
-            .store(RunState::Sending, Ordering::SeqCst);
-        self.schedule();
-    }
-
-    fn block_current_task_as_receiving(&self) -> ! {
-        let _guard = interrupt::uninterruptible();
-        let task = self.get_current_task().unwrap();
-        assert_eq!(
-            task.scheduler_state::<Self>().load(Ordering::SeqCst),
-            RunState::Running
-        );
-        task.scheduler_state::<Self>()
-            .store(RunState::Receiving, Ordering::SeqCst);
-        self.schedule();
-    }
 }
 
 pub type Scheduler = impl AbstractScheduler;
